// src/app/document/models/addNewfileAnnotate.model.ts
export class annotateFile {
    folder_id: string;
    user_id: number;
    project_id: string;
    parent_folder_id: any;
    folder_name: string;
    folder_level: number;
    is_folder_flag: boolean;
    folder_tag_name: string;
    folder_tag_id: number;
    is_hidden: boolean;
    current_project_id: string;
    old_folder_id: string;
    new_folder_id: string;
    new_document_name: string;
    new_folder_level: number;
    document_id: string;
    page_id: string;
    image_url: string;
    added_from_which_medium: string;
    document_name: string;
    created_date:string;
    last_updated_date:string;
    sync_version_uuid:string;
    updated_by_userid:string;
  }
  